Fix typo: rename Stretagy to Strategy
