+++ src/Strategy/FlyBehavior.java
@@ -1,4 +1,4 @@
-package Stretagy;
+package Strategy;
 
 public interface FlyBehavior {
