+++ src/Strategy/FlyNoWay.java
@@ -1,4 +1,4 @@
-package Stretagy;
+package Strategy;
 
 public class FlyNoWay implements FlyBehavior {
