+++ src/Strategy/FlyWithWings.java
@@ -1,4 +1,4 @@
-package Stretagy;
+package Strategy;
 
 public class FlyWithWings implements FlyBehavior {
