+++ src/Strategy/Quack.java
@@ -1,4 +1,4 @@
-package Stretagy;
+package Strategy;
 
 public class Quack implements QuackBehavior {
