+++ src/Strategy/Squeak.java
@@ -1,4 +1,4 @@
-package Stretagy;
+package Strategy;
 
 public class Squeak implements QuackBehavior {
